+++ extensions/python/src/main/java/com/hazelcast/jet/python/PythonTransforms.java
@@ -16,9 +16,9 @@
 package com.hazelcast.jet.python;
 
 import com.hazelcast.function.FunctionEx;
+import com.hazelcast.jet.annotation.EvolvingApi;
 import com.hazelcast.jet.pipeline.BatchStage;
 import com.hazelcast.jet.pipeline.StreamStage;
-import com.hazelcast.spi.annotation.Beta;
 
 import javax.annotation.Nonnull;
 
@@ -32,7 +32,7 @@ import javax.annotation.Nonnull;
  *
  * @since 4.0
  */
-@Beta
+@EvolvingApi
 public final class PythonTransforms {
 
     private PythonTransforms() {
