+++ hazelcast-jet-core/src/main/java/com/hazelcast/jet/pipeline/test/AssertionSinks.java
@@ -18,7 +18,6 @@ package com.hazelcast.jet.pipeline.test;
 
 import com.hazelcast.function.ConsumerEx;
 import com.hazelcast.jet.pipeline.Sink;
-import com.hazelcast.spi.annotation.Beta;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -43,7 +42,6 @@ import static java.util.concurrent.TimeUnit.SECONDS;
  *
  * @since 3.2
  */
-@Beta
 public final class AssertionSinks {
 
     private AssertionSinks() {
